Rename createTrxAwait to createTransaction
