+++ modules/database/server-ts/sql/index.ts
@@ -1,4 +1,4 @@
 export { default as knex } from './connector';
 export { default as populateTestDb } from './populateTestDb';
-export { default as createTrxAwait } from './createTrxAwait';
+export { default as createTransaction } from './createTransaction';
 export * from './helpers';
